+++ options.h
@@ -7,7 +7,7 @@
 int load_ini_file(FILE *);
 char *ini_get_string(char *section, char *key, const char *def);
 int ini_get_int(char *section, char *key, int def);
-double ini_get_double(char *section, char *key, int def);
+double ini_get_double(char *section, char *key, double def);
 int ini_get_bool(char *section, char *key, int def);
 void free_ini(void);
